"Fixed" buttons
Buttons weren't broken but there was a requirement for the head to be wearing the fnris "hat" before the brain could be seen. In a previous build I detactched the hat which was causing the button to not work unless you picked up the hat and put it on the head.

This requirement has been removed and now the 'activate' button